refactor: modularize i18n message structure
- Split messages into separate files by section (header, home, blog, etc.)
- Create modular type structure with individual type files
- Update language files to re-export from modular structure
- Improve maintainability and code organization
