+++ apps/docs/src/i18n/messages/en/blog.tsx
@@ -1,0 +1,14 @@
+import { BlogMessages } from "../types/blog";
+
+export const blog: BlogMessages = {
+  title: "SSGOI Blog",
+  description: "Check out the latest news, development stories, and tutorials from SSGOI.",
+  pageTitle: "SSGOI Blog",
+  pageDescription: "Latest news and development stories",
+  noPostsYet: "No posts yet.",
+  backToBlog: "← Back to Blog",
+  metadata: {
+    title: "SSGOI Blog - A New Experience in Web Page Transitions",
+    description: "Check out the latest news, development stories, and tutorials from SSGOI.",
+  },
+};
