+++ apps/docs/src/i18n/messages/en/console.tsx
@@ -1,0 +1,8 @@
+import { ConsoleMessages } from "../types/console";
+
+export const console: ConsoleMessages = {
+  welcome: "Welcome to SSGOI!",
+  subtitle: "Beautiful page transitions for modern web apps",
+  supports: "Supports React, Svelte, and more!",
+  visit: "Visit: https://github.com/meursyphus/ssgoi",
+};
