+++ apps/docs/src/i18n/messages/en/header.tsx
@@ -1,0 +1,14 @@
+import { HeaderMessages } from "../types/header";
+
+export const header: HeaderMessages = {
+  home: "Home",
+  blog: "Blog",
+  demo: "Demo",
+  github: "Github",
+  docs: "Docs",
+  tutorial: "Tutorial",
+  contributing: "Contributing",
+  showcase: "Showcase",
+  openMenu: "Open menu",
+  githubRepository: "GitHub Repository",
+};
