+++ apps/docs/src/i18n/messages/en/metadata.tsx
@@ -1,0 +1,31 @@
+import { MetadataMessages } from "../types/metadata";
+
+export const metadata: MetadataMessages = {
+  title: "SSGOI - Beautiful Page Transitions for Modern Web Apps",
+  description:
+    "SSGOI is a powerful page transition library that brings native app-like animations to the web. Create smooth, spring-based transitions with state preservation across all frameworks.",
+  keywords: [
+    "page transitions",
+    "animation library",
+    "react transitions",
+    "vue transitions",
+    "svelte transitions",
+    "spring animations",
+    "view transitions",
+    "web animations",
+    "ssgoi",
+  ],
+  og: {
+    title: "SSGOI - Beautiful Page Transitions for Modern Web Apps",
+    description:
+      "Create stunning page transitions with SSGOI. Native app-like animations, state preservation, and framework-agnostic design. Works with React, Vue, Svelte, and more.",
+    siteName: "SSGOI",
+    imageAlt: "SSGOI - Page Transition Library",
+  },
+  twitter: {
+    title: "SSGOI - Beautiful Page Transitions for Modern Web Apps",
+    description:
+      "Create stunning page transitions with SSGOI. Native app-like animations, state preservation, and framework-agnostic design.",
+    imageAlt: "SSGOI - Page Transition Library",
+  },
+};
