+++ apps/docs/src/i18n/messages/en/mobile-menu.tsx
@@ -1,0 +1,13 @@
+import { MobileMenuMessages } from "../types/mobile-menu";
+
+export const mobileMenu: MobileMenuMessages = {
+  documentToc: "Table of Contents",
+  menu: "Menu",
+  closeDrawer: "Close drawer",
+  noDocumentPage: "Please navigate to a documentation page.",
+  navigation: "Navigation",
+  documents: "Documentation",
+  blog: "Blog",
+  demo: "Demo",
+  language: "Language",
+};
