+++ apps/docs/src/i18n/messages/en/sidebar.tsx
@@ -1,0 +1,10 @@
+import { SidebarMessages } from "../types/sidebar";
+
+export const sidebar: SidebarMessages = {
+  categories: {
+    "getting-started": "Getting Started",
+    "core-concepts": "Core Concepts",
+    "view-transitions": "View Transitions",
+    "transitions": "Transitions",
+  },
+};
