+++ apps/docs/src/i18n/messages/ja/blog.tsx
@@ -1,0 +1,14 @@
+import { BlogMessages } from "../types/blog";
+
+export const blog: BlogMessages = {
+  title: "SSGOIブログ",
+  description: "SSGOIの最新ニュース、開発ストーリー、チュートリアルをご覧ください。",
+  pageTitle: "SSGOIブログ",
+  pageDescription: "最新ニュースと開発ストーリー",
+  noPostsYet: "まだ投稿がありません。",
+  backToBlog: "← ブログに戻る",
+  metadata: {
+    title: "SSGOIブログ - ウェブページトランジションの新しい体験",
+    description: "SSGOIの最新ニュース、開発ストーリー、チュートリアルをご覧ください。",
+  },
+};
