+++ apps/docs/src/i18n/messages/ja/console.tsx
@@ -1,0 +1,8 @@
+import { ConsoleMessages } from "../types/console";
+
+export const console: ConsoleMessages = {
+  welcome: "SSGOIへようこそ！",
+  subtitle: "モダンウェブアプリのための美しいページトランジション",
+  supports: "React、Svelteなど多様なフレームワークをサポート！",
+  visit: "訪問: https://github.com/meursyphus/ssgoi",
+};
