+++ apps/docs/src/i18n/messages/ja/header.tsx
@@ -1,0 +1,14 @@
+import { HeaderMessages } from "../types/header";
+
+export const header: HeaderMessages = {
+  home: "ホーム",
+  blog: "ブログ",
+  demo: "デモ",
+  github: "Github",
+  docs: "ドキュメント",
+  tutorial: "チュートリアル",
+  contributing: "貢献",
+  showcase: "ショーケース",
+  openMenu: "メニューを開く",
+  githubRepository: "GitHubリポジトリ",
+};
