+++ apps/docs/src/i18n/messages/ja/metadata.tsx
@@ -1,0 +1,31 @@
+import { MetadataMessages } from "../types/metadata";
+
+export const metadata: MetadataMessages = {
+  title: "SSGOI - モダンウェブアプリのための美しいページトランジション",
+  description:
+    "SSGOIは、ウェブにネイティブアプリのようなアニメーションをもたらす強力なページトランジションライブラリです。すべてのフレームワークで状態保持を備えたスムーズなスプリングベースのトランジションを作成します。",
+  keywords: [
+    "ページトランジション",
+    "アニメーションライブラリ",
+    "reactトランジション",
+    "vueトランジション",
+    "svelteトランジション",
+    "スプリングアニメーション",
+    "ビュートランジション",
+    "ウェブアニメーション",
+    "ssgoi",
+  ],
+  og: {
+    title: "SSGOI - モダンウェブアプリのための美しいページトランジション",
+    description:
+      "SSGOIで素晴らしいページトランジションを作成。ネイティブアプリのようなアニメーション、状態保持、フレームワーク非依存設計。React、Vue、Svelteなどで動作します。",
+    siteName: "SSGOI",
+    imageAlt: "SSGOI - ページトランジションライブラリ",
+  },
+  twitter: {
+    title: "SSGOI - モダンウェブアプリのための美しいページトランジション",
+    description:
+      "SSGOIで素晴らしいページトランジションを作成。ネイティブアプリのようなアニメーション、状態保持、フレームワーク非依存設計。",
+    imageAlt: "SSGOI - ページトランジションライブラリ",
+  },
+};
