+++ apps/docs/src/i18n/messages/ja/mobile-menu.tsx
@@ -1,0 +1,13 @@
+import { MobileMenuMessages } from "../types/mobile-menu";
+
+export const mobileMenu: MobileMenuMessages = {
+  documentToc: "目次",
+  menu: "メニュー",
+  closeDrawer: "閉じる",
+  noDocumentPage: "ドキュメントページに移動してください。",
+  navigation: "ナビゲーション",
+  documents: "ドキュメント",
+  blog: "ブログ",
+  demo: "デモ",
+  language: "言語 / Language",
+};
