+++ apps/docs/src/i18n/messages/ja/sidebar.tsx
@@ -1,0 +1,10 @@
+import { SidebarMessages } from "../types/sidebar";
+
+export const sidebar: SidebarMessages = {
+  categories: {
+    "getting-started": "はじめに",
+    "core-concepts": "コアコンセプト",
+    "view-transitions": "ビュートランジション",
+    "transitions": "トランジション",
+  },
+};
