+++ apps/docs/src/i18n/messages/ko/blog.tsx
@@ -1,0 +1,14 @@
+import { BlogMessages } from "../types/blog";
+
+export const blog: BlogMessages = {
+  title: "SSGOI 블로그",
+  description: "SSGOI의 최신 소식, 개발 스토리, 튜토리얼을 확인하세요.",
+  pageTitle: "SSGOI 블로그",
+  pageDescription: "최신 소식과 개발 스토리",
+  noPostsYet: "아직 게시물이 없습니다.",
+  backToBlog: "← 블로그로 돌아가기",
+  metadata: {
+    title: "SSGOI 블로그 - 웹 페이지 전환의 새로운 경험",
+    description: "SSGOI의 최신 소식, 개발 스토리, 튜토리얼을 확인하세요.",
+  },
+};
