+++ apps/docs/src/i18n/messages/ko/console.tsx
@@ -1,0 +1,8 @@
+import { ConsoleMessages } from "../types/console";
+
+export const console: ConsoleMessages = {
+  welcome: "SSGOI - 쓱오이에 오신 것을 환영합니다!",
+  subtitle: "모던 웹을 위한 아름다운 페이지 전환",
+  supports: "React, Svelte 등 다양한 프레임워크 지원!",
+  visit: "방문: https://github.com/meursyphus/ssgoi",
+};
