+++ apps/docs/src/i18n/messages/ko/header.tsx
@@ -1,0 +1,14 @@
+import { HeaderMessages } from "../types/header";
+
+export const header: HeaderMessages = {
+  home: "홈",
+  blog: "블로그",
+  demo: "데모",
+  github: "깃허브",
+  docs: "문서",
+  tutorial: "튜토리얼",
+  contributing: "기여하기",
+  showcase: "쇼케이스",
+  openMenu: "메뉴 열기",
+  githubRepository: "GitHub 저장소",
+};
