+++ apps/docs/src/i18n/messages/ko/index.tsx
@@ -1,0 +1,20 @@
+import { Messages } from "../types";
+import { header } from "./header";
+import { mobileMenu } from "./mobile-menu";
+import { home } from "./home";
+import { metadata } from "./metadata";
+import { sidebar } from "./sidebar";
+import { console } from "./console";
+import { blog } from "./blog";
+
+const ko: Messages = {
+  header,
+  mobileMenu,
+  home,
+  metadata,
+  sidebar,
+  console,
+  blog,
+};
+
+export default ko;
