+++ apps/docs/src/i18n/messages/ko/metadata.tsx
@@ -1,0 +1,19 @@
+import { MetadataMessages } from "../types/metadata";
+
+export const metadata: MetadataMessages = {
+  title: "SSGOI - 모던 웹을 위한 아름다운 페이지 전환 라이브러리",
+  description:
+    "SSGOI는 네이티브 앱과 같은 애니메이션을 웹에 구현하는 강력한 페이지 전환 라이브러리입니다. 상태 보존과 함께 부드러운 스프링 기반 전환을 모든 프레임워크에서 사용하세요.",
+  keywords: ["페이지 전환", "애니메이션 라이브러리", "리액트 전환", "뷰 전환", "스벨트 전환", "스프링 애니메이션", "뷰 트랜지션", "웹 애니메이션", "ssgoi"],
+  og: {
+    title: "SSGOI - 모던 웹을 위한 아름다운 페이지 전환 라이브러리",
+    description: "SSGOI로 멋진 페이지 전환을 만드세요. 네이티브 앱과 같은 애니메이션, 상태 보존, 프레임워크 독립적인 설계. React, Vue, Svelte 등 모든 프레임워크에서 작동합니다.",
+    siteName: "SSGOI",
+    imageAlt: "SSGOI - 페이지 전환 라이브러리",
+  },
+  twitter: {
+    title: "SSGOI - 모던 웹을 위한 아름다운 페이지 전환 라이브러리",
+    description: "SSGOI로 멋진 페이지 전환을 만드세요. 네이티브 앱과 같은 애니메이션, 상태 보존, 프레임워크 독립적인 설계.",
+    imageAlt: "SSGOI - 페이지 전환 라이브러리",
+  },
+};
