+++ apps/docs/src/i18n/messages/ko/mobile-menu.tsx
@@ -1,0 +1,13 @@
+import { MobileMenuMessages } from "../types/mobile-menu";
+
+export const mobileMenu: MobileMenuMessages = {
+  documentToc: "문서 목차",
+  menu: "메뉴",
+  closeDrawer: "닫기",
+  noDocumentPage: "문서 페이지로 이동해주세요.",
+  navigation: "네비게이션",
+  documents: "문서",
+  blog: "블로그",
+  demo: "데모",
+  language: "언어 / Language",
+};
