+++ apps/docs/src/i18n/messages/ko/sidebar.tsx
@@ -1,0 +1,10 @@
+import { SidebarMessages } from "../types/sidebar";
+
+export const sidebar: SidebarMessages = {
+  categories: {
+    "getting-started": "시작하기",
+    "core-concepts": "핵심 개념",
+    "view-transitions": "페이지 전환",
+    "transitions": "요소 전환",
+  },
+};
