+++ apps/docs/src/i18n/messages/types.ts
@@ -1,140 +1,2 @@
-import { ReactNode } from "react";
-
-export type Messages = {
-  header: {
-    home: string;
-    blog: string;
-    github: string;
-    docs: string;
-    tutorial: string;
-    contributing: string;
-    showcase: string;
-    openMenu: string;
-    githubRepository: string;
-    demo: string;
-  };
-  mobileMenu: {
-    documentToc: string;
-    menu: string;
-    closeDrawer: string;
-    noDocumentPage: string;
-    navigation: string;
-    documents: string;
-    blog: string;
-    language: string;
-    demo: string;
-  };
-  home: {
-    title: string;
-    getStarted: string;
-    readMore: string;
-    // Hero Section
-    badge: {
-      text: string;
-    };
-    heroTitle: {
-      line1: string;
-      line2: string;
-    };
-    subtitle: ReactNode;
-    description: string;
-    buttons: {
-      getStarted: string;
-      github: string;
-      demo: string;
-    };
-    quickInstall: {
-      react: string;
-      svelte: string;
-      vue: string;
-      solidjs: string;
-      qwik: string;
-    };
-    floatingBadges: {
-      performance: string;
-      stateMemory: string;
-    };
-    // Framework Support Section
-    frameworks: {
-      title: ReactNode;
-      subtitle: string;
-      comingSoon: string;
-    };
-    // Why SSGOI Section
-    whySSGOI: {
-      title: ReactNode;
-      subtitle: string;
-      features: {
-        ssr: {
-          title: string;
-          description: string;
-        };
-        browserCompat: {
-          title: string;
-          description: string;
-        };
-        zeroConfig: {
-          title: string;
-          description: string;
-        };
-      };
-    };
-    // Code Example Section
-    codeExample: {
-      title: ReactNode;
-      subtitle: string;
-    };
-    // CTA Section
-    cta: {
-      title: string;
-      subtitle: string;
-      buttons: {
-        viewDocs: string;
-        github: string;
-        demo: string;
-      };
-    };
-  };
-  metadata: {
-    title: string;
-    description: string;
-    keywords: string[];
-    og: {
-      title: string;
-      description: string;
-      siteName: string;
-      imageAlt: string;
-    };
-    twitter: {
-      title: string;
-      description: string;
-      imageAlt: string;
-    };
-  };
-  sidebar: {
-    categories: {
-      "getting-started": string;
-      "core-concepts": string;
-      "view-transitions": string;
-      "transitions": string;
-    };
-  };
-  console: {
-    welcome: string;
-    subtitle: string;
-    supports: string;
-    visit: string;
-  };
-  blog: {
-    title: string;
-    description: string;
-    pageTitle: string;
-    pageDescription: string;
-    noPostsYet: string;
-    backToBlog: string;
-    metadata: {
-      title: string;
-      description: string;
-    };
-  };
-};
+// Re-export from the new types directory
+export * from "./types";
