+++ apps/docs/src/i18n/messages/types/blog.ts
@@ -1,0 +1,12 @@
+export type BlogMessages = {
+  title: string;
+  description: string;
+  pageTitle: string;
+  pageDescription: string;
+  noPostsYet: string;
+  backToBlog: string;
+  metadata: {
+    title: string;
+    description: string;
+  };
+};
