+++ apps/docs/src/i18n/messages/types/console.ts
@@ -1,0 +1,6 @@
+export type ConsoleMessages = {
+  welcome: string;
+  subtitle: string;
+  supports: string;
+  visit: string;
+};
