+++ apps/docs/src/i18n/messages/types/header.ts
@@ -1,0 +1,12 @@
+export type HeaderMessages = {
+  home: string;
+  blog: string;
+  github: string;
+  docs: string;
+  tutorial: string;
+  contributing: string;
+  showcase: string;
+  openMenu: string;
+  githubRepository: string;
+  demo: string;
+};
