+++ apps/docs/src/i18n/messages/types/mobile-menu.ts
@@ -1,0 +1,11 @@
+export type MobileMenuMessages = {
+  documentToc: string;
+  menu: string;
+  closeDrawer: string;
+  noDocumentPage: string;
+  navigation: string;
+  documents: string;
+  blog: string;
+  language: string;
+  demo: string;
+};
