+++ apps/docs/src/i18n/messages/types/sidebar.ts
@@ -1,0 +1,8 @@
+export type SidebarMessages = {
+  categories: {
+    "getting-started": string;
+    "core-concepts": string;
+    "view-transitions": string;
+    "transitions": string;
+  };
+};
