+++ apps/docs/src/i18n/messages/zh/blog.tsx
@@ -1,0 +1,14 @@
+import { BlogMessages } from "../types/blog";
+
+export const blog: BlogMessages = {
+  title: "SSGOI博客",
+  description: "查看SSGOI的最新消息、开发故事和教程。",
+  pageTitle: "SSGOI博客",
+  pageDescription: "最新消息和开发故事",
+  noPostsYet: "暂无文章。",
+  backToBlog: "← 返回博客",
+  metadata: {
+    title: "SSGOI博客 - 网页过渡的全新体验",
+    description: "查看SSGOI的最新消息、开发故事和教程。",
+  },
+};
