+++ apps/docs/src/i18n/messages/zh/console.tsx
@@ -1,0 +1,8 @@
+import { ConsoleMessages } from "../types/console";
+
+export const console: ConsoleMessages = {
+  welcome: "欢迎使用SSGOI！",
+  subtitle: "为现代网页应用打造优美的页面过渡效果",
+  supports: "支持React、Svelte等多种框架！",
+  visit: "访问: https://github.com/meursyphus/ssgoi",
+};
