+++ apps/docs/src/i18n/messages/zh/header.tsx
@@ -1,0 +1,14 @@
+import { HeaderMessages } from "../types/header";
+
+export const header: HeaderMessages = {
+  home: "首页",
+  blog: "博客",
+  demo: "演示",
+  github: "Github",
+  docs: "文档",
+  tutorial: "教程",
+  contributing: "贡献",
+  showcase: "展示",
+  openMenu: "打开菜单",
+  githubRepository: "GitHub仓库",
+};
