+++ apps/docs/src/i18n/messages/zh/metadata.tsx
@@ -1,0 +1,31 @@
+import { MetadataMessages } from "../types/metadata";
+
+export const metadata: MetadataMessages = {
+  title: "SSGOI - 为现代网页应用打造优美的页面过渡效果",
+  description:
+    "SSGOI是一个强大的页面过渡库，为网页带来原生应用般的动画效果。创建流畅的基于弹簧的过渡效果，在所有框架中保持状态。",
+  keywords: [
+    "页面过渡",
+    "动画库",
+    "react过渡",
+    "vue过渡",
+    "svelte过渡",
+    "弹簧动画",
+    "视图过渡",
+    "网页动画",
+    "ssgoi",
+  ],
+  og: {
+    title: "SSGOI - 为现代网页应用打造优美的页面过渡效果",
+    description:
+      "使用SSGOI创建令人惊叹的页面过渡效果。原生应用般的动画、状态保持和框架无关设计。支持React、Vue、Svelte等。",
+    siteName: "SSGOI",
+    imageAlt: "SSGOI - 页面过渡库",
+  },
+  twitter: {
+    title: "SSGOI - 为现代网页应用打造优美的页面过渡效果",
+    description:
+      "使用SSGOI创建令人惊叹的页面过渡效果。原生应用般的动画、状态保持和框架无关设计。",
+    imageAlt: "SSGOI - 页面过渡库",
+  },
+};
