+++ apps/docs/src/i18n/messages/zh/mobile-menu.tsx
@@ -1,0 +1,13 @@
+import { MobileMenuMessages } from "../types/mobile-menu";
+
+export const mobileMenu: MobileMenuMessages = {
+  documentToc: "文档目录",
+  menu: "菜单",
+  closeDrawer: "关闭",
+  noDocumentPage: "请导航到文档页面。",
+  navigation: "导航",
+  documents: "文档",
+  blog: "博客",
+  demo: "演示",
+  language: "语言 / Language",
+};
