+++ apps/docs/src/i18n/messages/zh/sidebar.tsx
@@ -1,0 +1,10 @@
+import { SidebarMessages } from "../types/sidebar";
+
+export const sidebar: SidebarMessages = {
+  categories: {
+    "getting-started": "入门指南",
+    "core-concepts": "核心概念",
+    "view-transitions": "视图过渡",
+    "transitions": "过渡效果",
+  },
+};
